seoul-job-cnt/azure-func-connect/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
requests
pandas
urllib3
azure-eventhub>=5.10.0
azure-storage-blob>=12.20.0
python-dotenv>=1.0.0
websocket-client>=1.5.0
psycopg2-binary>=2.9.0